+++ app/Http/Controllers/DiamondColorController.php
@@ -7,6 +7,8 @@ use App\Traits\JsonResponse;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class DiamondColorController extends Controller
 {
@@ -20,11 +22,13 @@ class DiamondColorController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('diamond_color_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('diamond_color.index');
     }
 
     public function getData()
     {
+        abort_if(Gate::denies('diamond_color_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             return $this->diamond_color_service->getSource();
         } catch (Exception $e) {
@@ -34,6 +38,7 @@ class DiamondColorController extends Controller
 
     public function store(Request $request)
     {
+        abort_if(Gate::denies('diamond_color_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $validation = Validator::make(
             $request->all(),
@@ -80,6 +85,7 @@ class DiamondColorController extends Controller
 
     public function edit($id)
     {
+        abort_if(Gate::denies('diamond_color_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             return  $this->success(
                 config('enum.success'),
@@ -93,6 +99,7 @@ class DiamondColorController extends Controller
 
     public function status($id)
     {
+        abort_if(Gate::denies('diamond_color_status'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $diamond_color = $this->diamond_color_service->statusById($id);
             return $this->success(
@@ -107,6 +114,7 @@ class DiamondColorController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('diamond_color_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $diamond_color = $this->diamond_color_service->deleteById($id);
             return $this->success(
